ATENÇÃO: PROG. APENAS PARA FINS DE ESTUDO, NÃO FAÇA MAL USO DO MESMO.
In computing, a denial-of-service attack (DoS attack) is a cyber-attack
where the perpetrator seeks to make a machine or network resource
unavailable to its intended users by temporarily or indefinitely
disrupting services of a host connected to the Internet. Denial of
service is typically accomplished by flooding the targeted machine or
resource with superfluous requests in an attempt to overload systems and
prevent some or all legitimate requests from being fulfilled. A DoS
attack is analogous to a group of people crowding the entry door or gate
to a shop or business, and not letting legitimate parties enter into the
shop or business, disrupting normal operations.